Segment classes now associate themselves to their name in mission file.
