+++ src/fastoad/models/performances/mission/segments/hold.py
@@ -1,6 +1,6 @@
 """Class for simulating hold segment."""
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -15,7 +15,7 @@
 from .base import RegulatedThrustSegment, FixedDurationSegment
 
 
-class HoldSegment(RegulatedThrustSegment, FixedDurationSegment):
+class HoldSegment(RegulatedThrustSegment, FixedDurationSegment, mission_file_keyword="holding"):
     """
     Class for computing hold flight segment.
